Using the Box type to define your panels

It's inconvenient to have to modify the dimensions of panels depending on what type of joinery will be used. The Box type is intended to address this difficulty.

The outside dimensions of a Box are specified by its length, width and height.

By default a box has six faces. Each face is identified by an instance of a singleton type:

PanelCutting.Back — Type

The back face of a Box.

PanelCutting.Bottom — Type

The bottom face of a Box.

PanelCutting.Face — Type

Face

Face is the abstract supertype for the tokens that identify each face of a Box.

Subtypes are Top, Bottom, Left, Right, Front and Back, which are all singleton types.

PanelCutting.Front — Type

The front face of a Box.

PanelCutting.Left — Type

The left face of a Box.

PanelCutting.Right — Type

The right face of a Box.

PanelCutting.Top — Type

The top face of a Box.

For each face of the box, one can specify that that face is open/missing, or the material and thickness of stock from which to cut that face.

Grain Direction

For each face of a box, one can specify the grain direction.

PanelCutting.GDEither — Type

GDEither

Use GDEither() if you don't care whether the grain runs parallel to the long or short edges of the panel.

PanelCutting.GDLong — Type

GDLong

Use GDLong() to indicate that the grain of the face should run parallel to its longer edges.

PanelCutting.GDShort — Type

GDShort

Use GDShort() to indicate that the grain of the face should run parallel to its shorter edges.

PanelCutting.GrainDirection — Type

GrainDirection

GrainDirection is the abstract type for the choices of how the grain direction of the stock is oriented with respect to the shape of the panel.

Edges

Between each pair of adjacent faces is an Edge. For each edge, the type of joinery to be used at that edge can be specified.

PanelCutting.Edge — Type

Edge(::Face, ::Face)

The edge of the box that connects the two specified faces.

Joint Type

PanelCutting.ButtJoint — Type

ButtJoint(shortened)

If the joint type specified for a pair of Faces is a ButtJoint, then the face specified by shortened is shortened by the thickness of the other face.

PanelCutting.DadoJoint — Type

DadoJoint(tongued::Face, tongue_length::LengthType)

If the joint type specified for a pair of faces is a DadoJoint, then the face specified by tongued is shortened by the thickness of the other face and then lengthened by tongue_length.

PanelCutting.FingerJoint — Type

FingerJoint

If the joint type specified for a pair of faces is FingerJoint then the size of neither panel needs to be adjusted for that joint.

If you intend to make blind finger or dovetail joints, use DadoJoint.

If you want a miter joint, use FingerJoint.

PanelCutting.JointType — Type

JointType

The abstract supertype for objects describing the joint type bwtween two faces.

Defining a box

A box is first specified by its outside dimensions. The Unitful package (and optionally UnitfulUS) are used so we can specify units for the box dimensions.

PanelCutting.Box — Type

Box(length::LengthType, width::LengthType, height::LengthType)

Define a box of the spcified dimensions.

using PanelCutting, Unitful

my_box = Box(0.3u"m", 0.2u"m", 0.1u"m")

Box(0.3 m, 0.2 m, 0.1 m, DataStructures.DefaultDict{Any, Any, Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(inch,), 𝐋, nothing}}}(), DataStructures.DefaultDict{Any, Any, Nothing}(), DataStructures.DefaultDict{Any, Any, Bool}(), DataStructures.DefaultDict{Any, Any, GDLong}(), DataStructures.DefaultDict{Any, Any, FingerJoint}())

Once the box is defined, one can specify the thickness and material of each face, whether a face is open, and its grain direction:

my_box.open[Top()] = true

do_faces(my_box) do face
    # All faces are 5mm baltic birch:
    my_box.thickness[face] = 5u"mm"
    my_box.material[face] = "Baltic Birch"
end

One can also specify the joint type to be used for each edge:

let
    # Use finger joints between each pair of sides:
    sides = [Left(), Back(), Right(), Front()]
    for i in 0:(length(sides) - 1)
        e = Edge(sides[i + 1],
                 sides[1 + (i + 1) % length(sides)])
        my_box.joint_types[e] = FingerJoint()
    end
    # Use a dado joint for each edge of the bottom:
    for n in neighbors(Bottom())
        edge = Edge(Bottom(), n)
        my_box.joint_types[edge] = DadoJoint(Bottom(), 2u"mm")
    end
end

Once the box is fully specified, we can ask for its WantedPanels:

WantedPanels(my_box)

5-element Vector{AbstractWantedPanel}:
 WantedPanel(1, 0.2 m, 0.1 m, 5 mm, "Baltic Birch", "Back")
 WantedPanel(2, 0.294 m, 0.194 m, 5 mm, "Baltic Birch", "Bottom")
 WantedPanel(3, 0.2 m, 0.1 m, 5 mm, "Baltic Birch", "Front")
 WantedPanel(4, 0.3 m, 0.1 m, 5 mm, "Baltic Birch", "Left")
 WantedPanel(5, 0.3 m, 0.1 m, 5 mm, "Baltic Birch", "Right")

Related Functions

PanelCutting.distance — Function

distance(::Box, ::Face, ::Face)

Return the outside dimension of the box between the two faces.

PanelCutting.do_faces — Function

do_faces(::Function, ::Box)

Apply the function to each face of the Box that is not open/missing.

PanelCutting.neighbors — Function

neighbors(face1::Face, face2::Face)::Boll

Return true if face1 is a neighbor of face2.

neighbors(face::Face)

Return the Faces that are adjacent to the specified Face.

PanelCutting.opposite — Function

opposite(::Face)::Face

Return the Face that is opposite to the specified face.

PanelCutting.WantedPanels — Function

WantedPanels(::Box)

Return the collection of WantedPanels to make the Box.